refactor: moved utility functions to internal
